chore(tests): fix + move relocation test
